I mistakenly had called ByteBuffer.getInt() in a number of place in Ensemble.java when there were only 2 bytes of data in the buffer.  I've changed this call to getShort().
git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@162 9e959af2-1677-4b95-8350-622619c7fe3b

+++ src/java/edu/hawaii/soest/kilonalu/adcp/Ensemble.java
@@ -346,9 +346,9 @@ public class Ensemble {
     * A method that returns the Ensemble headerID field contents 
     * as a int.
     */
-   public int getHeaderID(){
+   public short getHeaderID(){
      return ensembleHeader.getHeaderID().order(
-       ByteOrder.LITTLE_ENDIAN).getInt();
+       ByteOrder.LITTLE_ENDIAN).getShort();
    }
 
    /**
